test: cleans up fields-relationship test suite
