+++ test/fields-relationship/baseFields.ts
@@ -1,0 +1,8 @@
+import type { CollectionConfig } from 'payload'
+
+export const baseRelationshipFields: CollectionConfig['fields'] = [
+  {
+    name: 'name',
+    type: 'text',
+  },
+]
