+++ test/fields-relationship/collections/FilterTrue/index.ts
@@ -1,0 +1,12 @@
+import type { CollectionConfig } from 'payload'
+
+import { baseRelationshipFields } from '../../baseFields.js'
+import { relationTrueFilterOptionSlug } from '../../slugs.js'
+
+export const RelationshipFilterTrue: CollectionConfig = {
+  admin: {
+    useAsTitle: 'name',
+  },
+  fields: baseRelationshipFields,
+  slug: relationTrueFilterOptionSlug,
+}
