+++ test/fields-relationship/collections/MixedMedia/index.ts
@@ -1,0 +1,19 @@
+import type { CollectionConfig } from 'payload'
+
+import {
+  mixedMediaCollectionSlug,
+  podcastCollectionSlug,
+  videoCollectionSlug,
+} from '../../slugs.js'
+
+export const MixedMedia: CollectionConfig = {
+  slug: mixedMediaCollectionSlug,
+  fields: [
+    {
+      type: 'relationship',
+      name: 'relatedMedia',
+      relationTo: [videoCollectionSlug, podcastCollectionSlug],
+      hasMany: true,
+    },
+  ],
+}
