+++ test/fields-relationship/collections/Relation1/index.ts
@@ -1,0 +1,9 @@
+import type { CollectionConfig } from 'payload'
+
+import { baseRelationshipFields } from '../../baseFields.js'
+import { relationOneSlug } from '../../slugs.js'
+
+export const Relation1: CollectionConfig = {
+  fields: baseRelationshipFields,
+  slug: relationOneSlug,
+}
