+++ test/fields-relationship/collections/Relation2/index.ts
@@ -1,0 +1,9 @@
+import type { CollectionConfig } from 'payload'
+
+import { baseRelationshipFields } from '../../baseFields.js'
+import { relationTwoSlug } from '../../slugs.js'
+
+export const Relation2: CollectionConfig = {
+  fields: baseRelationshipFields,
+  slug: relationTwoSlug,
+}
