+++ test/fields-relationship/slugs.ts
@@ -13,3 +13,19 @@ export const videoCollectionSlug = 'videos'
 export const podcastCollectionSlug = 'podcasts'
 export const mixedMediaCollectionSlug = 'mixed-media'
 export const versionedRelationshipFieldSlug = 'versioned-relationship-field'
+
+export const collectionSlugs = [
+  relationOneSlug,
+  relationTrueFilterOptionSlug,
+  relationFalseFilterOptionSlug,
+  relationTwoSlug,
+  relationRestrictedSlug,
+  relationWithTitleSlug,
+  relationUpdatedExternallySlug,
+  collection1Slug,
+  collection2Slug,
+  videoCollectionSlug,
+  podcastCollectionSlug,
+  mixedMediaCollectionSlug,
+  versionedRelationshipFieldSlug,
+]
